+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- GSEA now has score and esScore so we can give two scores to a gene, its original score than then the running enrichment score it is assigned.
 - Revised argsort. Refactored ext gsea. Viper now sorts genesets even if unneccessary.
 - Refined names for gsea and ext gsea to be less confusing.
 - ext gsea shares more code with original gsea plot. Ext gsea now plots the full es for the curve to make line smoother.